Edit model, controller, and loop fixtures to clean up after all tests run.
There was a bug where all tests would fail if PYTEST_KEEP_MODELS was not set,
as the first test would start cleaning up the models since it would return to
run the logic after yield. Setting autouse=True and scope to session allows
this cleanup to happen after all tests in the session are run (which is ideal
for these types of tests).