Add utilities for running stuff on units.
As all async jujulib commands need to be run in a loop.run() wrapper, it makes
things ugly and hard to unit test. This commit adds utility functions to do these
calls to allow for better testing for code that uses these async calls.

Also, juju actions now are checked for "completed" status; if other status,
ActionError is raised.