Add max_frame_size as param for juju.Model creation.
There's an issue with juju run and juju run-action commands that
can cause connection issues if the parameter/result is too large [0].
This issue manifests in python-libjuju when running the tool in a large environment [1].
Setting the max_frame_size when instantiating juju.Model helps combat this issue.
When LP#1896209 is fixed, we can look at removing max_frame_size.

[0] https://bugs.launchpad.net/juju/+bug/1896209
[1] https://github.com/juju/python-libjuju/issues/523