Implement backup tracking and report output.
This commit adds a json formatted output of the backups that succeeded
and failed. When processing backups, the resulting entries are added
to the BackupTracker. The tracker then prints the json results after
all backups have been processed.

To help accomplish this, backup actions now return the path the backup file
was saved to.