Refactor linting tool configurations.
In an effort to centralize configurations, this commit moves
isort and black config to the new pyproject.toml. Since flake8
does not support configs in that file, its config is moved to a
dedicated .flake8 file.

This commit also moves the bulk of setup.py configuration to the
best-practice-preferred setup.cfg file.

tox has also been simplified to only have provide a "lint" and a
"format" environment. These envs share the "lintbase" dependencies
and do the same linting/auto-formatting as before (black, flake8,
isort).