Change backup_controllers to use py-libjuju call.
Previously, we had to backup controllers by using subprocess and
running a `juju backup-model` command as the library did not provide
that functionality. This functionality has been added per GitHub
Issue #528. This commit added various model backup functions.

Unfortunately, there is not yet a way to download the backup directly
using the library. This requires us to change the permissions of the tmp
parent dir containing the backup, pulling down the file, and then removing
the parent directory. Various machine-based utils have been added to utils.py
for assisting in those functions.

A lot of this logic in backup_controller can be removed when python-libjuju
provides functionality for also downloading backups from the controller.

Related-Bug: LP#1941918