Change snap confinement from classic to strict.
Now that there are no subprocess calls being made, this tool can
now be a confined snap. This allows built snaps to be pushed directly
to the snap store without a review.

In this commit, we change the confinement to strict; to accomplish
this, we add required plugs to juju-backup-all app as well as add
openssh-client package for use of ssh/scp binaries.

Because we can now run in a confined snap, we need to check if we're
running in a snap environment; if so, set JUJU_DATA accordingly
(using SNAP_REAL_HOME) as os.path.expanduser() uses the snap home
rather than the real home of the user, so it's not a reliable way
of finding the juju config.

Closes-Bug: LP#1941919