Add handling in the case of no unit being a leader.
In Juju, there are rare cases where no unit is elected a leader.
This can occur if all units are in a lost state. This would cause
our get_leader function to return None, which would break the
proceeding calls using that leader unit.

To combat this, we introduce error handling for the case of no
leader. get_leader has been enhanced to raise a NoLeaderError if
no leader is returned. Additionally, the error bubbles up to
process.py which records the error and gracefully continues with
the rest of the apps.