Refactor backup_app exception handling into one except block.
There was a lot of duplicated code for handling the different possible errors
during backup_app execution. Since the flow is basically the same for each
(i.e. log the error and add the error to the tracker), we can simplify this
by having the error classes give more info and making the _log and add_error
calls more generic.